# swagger.yaml
openapi: 3.0.0
info:
  title: Node MySQL API
  description: API for user authentication and account management
  version: 1.0.0
servers:
  - url: http://localhost:4000
    description: Development server

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

paths:
  /accounts/register:
    post:
      summary: Register a new account
      tags: [Accounts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, firstName, lastName, email, password, confirmPassword, acceptTerms]
              properties:
                title:
                  type: string
                  example: Mr
                firstName:
                  type: string
                  example: John
                lastName:
                  type: string
                  example: Doe
                email:
                  type: string
                  example: john@example.com
                password:
                  type: string
                  minLength: 6
                  example: password123
                confirmPassword:
                  type: string
                  example: password123
                acceptTerms:
                  type: boolean
                  example: true
      responses:
        '200':
          description: Registration successful, please check your email for verification instructions
        '400':
          description: Validation error

  /accounts/authenticate:
    post:
      summary: Authenticate account and get JWT token
      tags: [Accounts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  example: john@example.com
                password:
                  type: string
                  example: password123
      responses:
        '200':
          description: Authentication successful, returns JWT token
        '400':
          description: Invalid credentials

  /accounts/refresh-token:
    post:
      summary: Use a refresh token to generate a new JWT token and a new refresh token
      tags: [Accounts]
      parameters:
        - in: cookie
          name: refreshToken
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Token refreshed successfully
        '401':
          description: Invalid token

  /accounts/revoke-token:
    post:
      summary: Revoke a refresh token
      tags: [Accounts]
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
      responses:
        '200':
          description: Token revoked
        '400':
          description: Token is required
        '401':
          description: Unauthorized

  /accounts/verify-email:
    post:
      summary: Verify a new account with a verification token received by email
      tags: [Accounts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token:
                  type: string
      responses:
        '200':
          description: Verification successful, you can now login
        '400':
          description: Invalid token

  /accounts/forgot-password:
    post:
      summary: Submit email address to reset the password on an account
      tags: [Accounts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  example: john@example.com
      responses:
        '200':
          description: Please check your email for password reset instructions

  /accounts/validate-reset-token:
    post:
      summary: Validate the reset password token received by email
      tags: [Accounts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token:
                  type: string
      responses:
        '200':
          description: Token is valid
        '400':
          description: Invalid token

  /accounts/reset-password:
    post:
      summary: Reset the password on an account
      tags: [Accounts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, password, confirmPassword]
              properties:
                token:
                  type: string
                password:
                  type: string
                  minLength: 6
                  example: newpassword123
                confirmPassword:
                  type: string
                  example: newpassword123
      responses:
        '200':
          description: Password reset successful, you can now login
        '400':
          description: Invalid token

  /accounts:
    get:
      summary: Get all accounts (Admin only)
      tags: [Accounts]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of all accounts
        '401':
          description: Unauthorized
    post:
      summary: Create a new account (Admin only)
      tags: [Accounts]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, firstName, lastName, email, password, confirmPassword, role]
              properties:
                title:
                  type: string
                  example: Mr
                firstName:
                  type: string
                  example: John
                lastName:
                  type: string
                  example: Doe
                email:
                  type: string
                  example: john@example.com
                password:
                  type: string
                  minLength: 6
                  example: password123
                confirmPassword:
                  type: string
                  example: password123
                role:
                  type: string
                  enum: [Admin, User]
                  example: User
      responses:
        '200':
          description: Account created successfully
        '400':
          description: Validation error

  /accounts/{id}:
    get:
      summary: Get account by ID
      tags: [Accounts]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Account found
        '401':
          description: Unauthorized
        '404':
          description: Account not found
    put:
      summary: Update an account
      tags: [Accounts]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                firstName:
                  type: string
                lastName:
                  type: string
                email:
                  type: string
                password:
                  type: string
                confirmPassword:
                  type: string
                role:
                  type: string
                  enum: [Admin, User]
      responses:
        '200':
          description: Account updated successfully
        '401':
          description: Unauthorized
    delete:
      summary: Delete an account
      tags: [Accounts]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Account deleted successfully
        '401':
          description: Unauthorized